CAMEL-13313: Add support for generating type converter loader source code to be able to load component type converters in a faster way
