+++ tooling/apt/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -21,5 +21,5 @@ org.apache.camel.tools.apt.EndpointAnnotationProcessor
 org.apache.camel.tools.apt.SpiProcessor
 org.apache.camel.tools.apt.CoreTypeConverterProcessor
 org.apache.camel.tools.apt.TypeConverterProcessor
-org.apache.camel.tools.apt.TypeConverterLoaderProcessor
+org.apache.camel.tools.apt.TypeConverterLoaderGenerator
